Move components to specific directories, split the main App component into several ones
