+++ desktop-web-app/web-app/src/components/account/editable-avatar.tsx
