+++ desktop-web-app/web-app/src/components/appearance/theme-provider.tsx
@@ -22,7 +22,7 @@
 
 import React, {createContext, useContext, useEffect, useState} from "react";
 import {getAppearance} from "@/rpc/app-preferences-service.ts";
-import {fromTheme, systemTheme, ThemeOption} from "@/components/converter/theme.ts";
+import {fromTheme, systemTheme, ThemeOption} from "@/converter/theme.ts";
 import {saveThemeInStorage, themeFromStorage} from "@/storage/appearance.ts";
 
 /**
